Dev: move ChartConfig Indicator Explorer tests to that folder. Start test for chart config saving
