+++ explorer/indicatorExplorer/ExplorableChartConfig.jsdom.test.ts
