+++ explorer/indicatorExplorer/IndicatorUtils.ts
